+++ Imagekit.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.1.32319.34
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.32407.337
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Imagekit", "Imagekit\Imagekit.csproj", "{D7929DCF-13EC-4677-9859-D55DA9D752E4}"
 EndProject
@@ -11,8 +11,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Imagekit.UnitTests", "Imagekit.UnitTests\Imagekit.UnitTests.csproj", "{B459E977-A7C2-47CB-A743-2F0AAFCE8178}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ImagekitSample", "ImagekitSample\ImagekitSample.csproj", "{06ED9104-1599-4956-BCA6-0C8D4F9743BF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -28,10 +26,6 @@ Global
 		{B459E977-A7C2-47CB-A743-2F0AAFCE8178}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B459E977-A7C2-47CB-A743-2F0AAFCE8178}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B459E977-A7C2-47CB-A743-2F0AAFCE8178}.Release|Any CPU.Build.0 = Release|Any CPU
-		{06ED9104-1599-4956-BCA6-0C8D4F9743BF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{06ED9104-1599-4956-BCA6-0C8D4F9743BF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{06ED9104-1599-4956-BCA6-0C8D4F9743BF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{06ED9104-1599-4956-BCA6-0C8D4F9743BF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
